🔨 fix: fixed various bugs

+++ components/shared/navigation/FileTree/Items/DialogConfirmDeleteItem.tsx
@@ -1,24 +1,43 @@
-import { useState } from "react";
+import { useState, useEffect } from "react";
 
 import { Flex, Text, Button, Dialog, Icon } from "components/shared";
 import { CrossIcon, FolderIcon, FileIcon } from "components/icons";
-import { useFiles, useFileAttachments } from "hooks";
+import { useFiles, useFileAttachments, useToast } from "hooks";
 
 import { Action } from "./Action";
 
 export const DialogConfirmDeleteItem = ({ item, onRemove }) => {
+  const { setToast } = useToast();
   const [dialogOpen, setDialogOpen] = useState(false);
   const { setSelectedFile } = useFiles();
-  const { deleteFileAttachment } = useFileAttachments();
+  const { deleteFileAttachment, deleteFileAttachmentResult } =
+    useFileAttachments();
+
   const handleRemoveItem = async () => {
     const deletedFile = await deleteFileAttachment({
       where: {
         id: item.id,
       },
     });
-    setSelectedFile(null);
-    onRemove();
   };
+
+  useEffect(() => {
+    if (Boolean(deleteFileAttachmentResult.data)) {
+      setToast({
+        title: "Success",
+        description: `Module ${item?.label} has been deleted`,
+        color: "success",
+      });
+      setSelectedFile(null);
+      onRemove();
+    } else if (Boolean(deleteFileAttachmentResult.error)) {
+      setToast({
+        title: "Error",
+        description: `Something went wrong with deleting ${item?.label}`,
+        color: "error",
+      });
+    }
+  }, [deleteFileAttachmentResult]);
 
   return (
     <Dialog open={dialogOpen} onOpenChange={setDialogOpen}>
@@ -72,6 +91,7 @@ export const DialogConfirmDeleteItem = ({ item, onRemove }) => {
             <Button color="secondary">{`Cancel`}</Button>
           </Dialog.Close>
           <Button
+            state={deleteFileAttachmentResult.fetching ? "loading" : "default"}
             autoFocus
             onClick={() => handleRemoveItem()}
             color="error"
